refactor getBookHighlights
