+++ src/services/BookHighlightService.ts
@@ -1,12 +1,13 @@
 'use server';
 
-import BookHighlight from '@/types/BookHighlight';
+import BookHighlight, { BookHighlightPreview } from '@/types/BookHighlight';
 import { Result } from '@/types/Result';
 import bookHighlightsRepository from './repositories/BookHighlightsRepository';
 import { revalidatePath } from 'next/cache';
+import { getUserEmail, isAuthenticated } from './UserService';
 
 export async function addBookHighlight(
-  highlightData: BookHighlight
+  highlightData: BookHighlight,
 ): Promise<Result<string>> {
   try {
     await bookHighlightsRepository.createHighlight(highlightData);
@@ -18,7 +19,7 @@ export async function addBookHighlight(
 }
 
 export async function deleteBookHighlight(
-  highlightId: number
+  highlightId: number,
 ): Promise<Result<string>> {
   try {
     await bookHighlightsRepository.deleteHighlight(highlightId);
@@ -27,5 +28,23 @@ export async function deleteBookHighlight(
   } catch (error) {
     console.log(error);
     return { success: false, error: 'Failed to delete book highlight' };
+  }
+}
+
+export async function getBookHighlights(
+  googleId: string,
+): Promise<Result<BookHighlightPreview[]>> {
+  try {
+    const userEmail = await getUserEmail();
+    await isAuthenticated(userEmail);
+
+    const highlights = await bookHighlightsRepository.getHighlights(
+      googleId,
+      userEmail,
+    );
+    return { success: true, data: highlights };
+  } catch (error) {
+    console.error('Error getting book highlights:', error);
+    return { success: false, error: 'Failed to get book highlights' };
   }
 }
