+++ src/services/repositories/BookHighlightsRepository.ts
@@ -1,5 +1,6 @@
-import BookHighlight from '@/types/BookHighlight';
+import BookHighlight, { BookHighlightPreview } from '@/types/BookHighlight';
 import { db } from '../database/kysely';
+import { highlightAdapter } from '../adapters/BookAdapter';
 
 class BookHighlightsRepository {
   async createHighlight(bookHighlight: BookHighlight): Promise<void> {
@@ -23,6 +24,34 @@ class BookHighlightsRepository {
       .deleteFrom('readit_books_highlights')
       .where('id', '=', highlightId)
       .execute();
+  }
+
+  async getHighlights(
+    googleId: string,
+    userEmail: string,
+  ): Promise<BookHighlightPreview[]> {
+    const result = await db
+      .selectFrom('readit_books as rb')
+      .fullJoin(
+        'readit_books_highlights as rbh',
+        'rb.book_hash',
+        'rbh.book_hash',
+      )
+      .select([
+        'rb.id as book_id',
+        'rbh.id as highlight_id',
+        'rb.title',
+        'rb.authors',
+        'rbh.highlight_text',
+        'rbh.page',
+        'rbh.created_at',
+      ])
+      .where('rb.google_id', '=', googleId)
+      .where('rb.user_email', '=', userEmail)
+      .orderBy('rbh.created_at', 'asc')
+      .execute();
+
+    return highlightAdapter(result);
   }
 }
